requirements.txt
google-generativeai
streamlit
diffusers